+++ sources/EngineWorks/DBAL/Traits/ResultImplementsCountable.php
@@ -1,0 +1,10 @@
+<?php
+namespace EngineWorks\DBAL\Traits;
+
+trait ResultImplementsCountable
+{
+    public function count()
+    {
+        return $this->resultCount();
+    }
+}