+++ sources/EngineWorks/DBAL/Traits/ResultImplementsIterator.php
@@ -1,0 +1,12 @@
+<?php
+namespace EngineWorks\DBAL\Traits;
+
+use EngineWorks\DBAL\Iterators\ResultIterator;
+
+trait ResultImplementsIterator
+{
+    public function getIterator()
+    {
+        return new ResultIterator($this);
+    }
+}